Move around the content of connection.ts
